+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Code coverage. [#11](https://github.com/tzaeschke/phtree-cpp/pull/11)
-- Global JUnit callback for initial setup. THis allows setting global properties before cantrally
+- Global JUnit callback for initial setup. THis allows setting global properties before centrally
   before running any tests.
   [#38](https://github.com/netsec-ethz/scion-java-client/pull/38)
- 
-  
+- Support for `/etc/scion/hosts` and for OS search domains (e.g. `/etc/resolv.conf`). 
+  [#40](https://github.com/netsec-ethz/scion-java-client/pull/40)
+
 ### Changed
 - BREAKING CHANGE: Changed maven artifactId to "client"
   [#9](https://github.com/netsec-ethz/scion-java-client/pull/9)